- Se solucionan los problemas de escritura del HDF5 cuando la configuracion es muy grande - Se sustituye el tipo de dato para las dimensiones del reticulo de uint a size_t, para admitir grandes configuraciones

+++ axions/include/index.h
@@ -5,9 +5,9 @@ namespace indexXeon
 #ifdef	USE_XEON
 	__attribute__((target(mic)))
 #endif
-	static inline void idx2Vec(uint idx, uint x[3], const uint Lx)
+	static inline void idx2Vec(size_t idx, size_t x[3], const size_t Lx)
 	{
-			uint tmp = idx/Lx;
+			size_t tmp = idx/Lx;
 
 			x[2] = tmp/Lx;
 			x[1] = tmp - x[2]*Lx;
@@ -17,7 +17,7 @@ namespace indexXeon
 #ifdef	USE_XEON
 	__attribute__((target(mic)))
 #endif
-	static inline uint vec2Idx(uint x[3], const uint Lx)
+	static inline size_t vec2Idx(size_t x[3], const size_t Lx)
 	{
 		return (x[0] + Lx*(x[1] + Lx*x[2]));
 	}
